simplify options for rhp
(cherry picked from commit 268cd320f3ff28d5481004f6fa04551675bfc1b0)
